For messages without a date, use 2000-01-01 instead of the current date
This applies to the GGA, GLL, GNS and AIVDM type 5 messages.

This adheres better to the principle of least surprise for those that are
unaware that GGA, GLL, GNS message do not encode the date.

It removes the following edge cases:
* Processing the same recorded data gives different results depending on the
current date
* Messages from the same source giving different dates, depending on the
configuration of the local machine (reported in #24)

Of course, there is an initial surprise when these messages return a date of
2000-01-01 but the user easily learns that these are invalid and that they
should use (for example) the RMC message instead.

This also allows the dependency of the chrono crate on `std` to be removed,
which was the final item preventing this crate from being used in no_std
environments.